requirements.txt

cython
numpy
flask
pandas
pytorch 
torchvision


